Use govuk-style markup for default table block HTML
Gutenberg has a "table" block provided by default, but the HTML it produces does not pick up the styling from the GOVUK Frontend, as it has none of the classes applied.

We therefore had a couple of options:
1. Amend our CSS to style the HTML the block outputs by default to look like a GOVUK-style table
2. Amend the HTML output from the table block, so it includes all the govuk-prefixed classes, and therefore picks up the correct styling.

We've opted for option 2, as it's probably more likely that the CSS rules for those classes will be tweaked in future, rather than that the class names themselves will change. We should therefore be able to e.g. pick up styling tweaks from minor version bumps in the GOVUK Frontend, without having to recompile our own CSS.

There is a risk here, in that if the default HTML output of the Gutenberg table block changes in future (e.g. by adding classes to the table element) our code would not longer identify the elements that should be replaced. So we might want to refactor this in future, to be more flexible (e.g. by doing regular expression replacements, rather than using str_replace(). But this should work for now.

Resolves: #6
Resolves: https://trello.com/c/OTn4G0b4/29-styling-for-tables