Do not print tags directly into the header
Because the child theme always loads before the parent theme, there is no way of over-riding this function, which means we'd always end up with a couple of GOV.UK icons in the header, and an unnecessary meta tag.

These aren't required in this theme as a standalone theme, and it should be up to child themes to decide what image assets/meta tags they want in the header.