Add search box to default header
The govuk design system doesn't have a header component that includes a
search form, and the style used on gov.uk (where search is a massive
dropdown box) isn't appropriate for most of our sites, so I've had to
just put together what seems like a sensible approach here (although the
HTML amd css is still largely based on the gov.uk search box).

The search box lives in the top right on tablet and desktop, and drops
below the site title and strapline on mobile, becoming full width.