Add aspect-ratio helper mixin
This mixin leverages CSS to constrain display of an image to a defined
aspect ratio. Images are zoomed and overlay hidden to ensure the image
fills the desired area. This helps with situations where the content
editor has uploaded a non-standard image size and we want to ensure
consistent display in a component.

We can also change the aspect ratio across breakpoints to enable a
degree of art direction using the same image.

Example usage that would give a 1.5:1 aspect-ratio:

```
.image-wrapper {
  @include aspect-ratio(600, 400);
}
```